Shutdown handling refactor improvement
* refactor: move nomos-utils/lifecycle.rs.rs->nomos-services/utils'

* refactor: rename `nomos-services/utils` crate to `services-utils`
