+++ nomos-services/blend/Cargo.toml
@@ -16,7 +16,7 @@ nomos-blend-network = { path = "../../nomos-blend/network", features = [
 ] }
 nomos-blend-message = { path = "../../nomos-blend/message" }
 nomos-network = { path = "../network" }
-nomos-utils = { path = "../../nomos-utils" }
+services-utils = { path = "../utils" }
 overwatch-rs = { git = "https://github.com/logos-co/Overwatch", rev = "f5f7ea0" }
 rand = "0.8.5"
 rand_chacha = "0.3"
