+++ nomos-services/data-availability/indexer/src/lib.rs
@@ -18,7 +18,6 @@ use nomos_mempool::{backend::MemPool, network::NetworkAdapter as MempoolAdapter}
 use nomos_storage::backends::StorageBackend;
 use nomos_storage::StorageService;
 use nomos_tracing::info_with_id;
-use nomos_utils::lifecycle;
 use overwatch_rs::services::handle::ServiceStateHandle;
 use overwatch_rs::services::relay::{Relay, RelayMessage};
 use overwatch_rs::services::state::{NoOperator, NoState};
@@ -27,6 +26,7 @@ use overwatch_rs::DynError;
 use rand::{RngCore, SeedableRng};
 use serde::de::DeserializeOwned;
 use serde::{Deserialize, Serialize};
+use services_utils::overwatch::lifecycle;
 use storage::DaStorageAdapter;
 use tokio::sync::oneshot::Sender;
 use tracing::instrument;
