+++ nomos-services/data-availability/network/Cargo.toml
@@ -13,7 +13,7 @@ nomos-core = { path = "../../../nomos-core/chain-defs" }
 nomos-libp2p = { path = "../../../nomos-libp2p" }
 nomos-da-network-core = { path = "../../../nomos-da/network/core" }
 nomos-tracing = { path = "../../../nomos-tracing" }
-nomos-utils = { path = "../../../nomos-utils" }
+services-utils = { path = "../../utils" }
 subnetworks-assignations = { path = "../../../nomos-da/network/subnetworks-assignations" }
 libp2p = { version = "0.53", features = ["ed25519"] }
 serde = { version = "1.0", features = ["derive"] }
