+++ nomos-services/data-availability/verifier/Cargo.toml
@@ -18,7 +18,7 @@ nomos-da-network-core = { path = "../../../nomos-da/network/core" }
 nomos-da-network-service = { path = "../../../nomos-services/data-availability/network" }
 nomos-storage = { path = "../../../nomos-services/storage" }
 nomos-tracing = { path = "../../../nomos-tracing" }
-nomos-utils = { path = "../../../nomos-utils" }
+services-utils = { path = "../../utils" }
 overwatch-rs = { git = "https://github.com/logos-co/Overwatch", rev = "f5f7ea0" }
 overwatch-derive = { git = "https://github.com/logos-co/Overwatch", rev = "f5f7ea0" }
 serde = { version = "1.0", features = ["derive"] }
