+++ nomos-services/key-management-system/src/lib.rs
@@ -3,12 +3,12 @@ use crate::secure_key::SecuredKey;
 use bytes::Bytes;
 use futures::StreamExt;
 use log::error;
-use nomos_utils::lifecycle;
 use overwatch_rs::services::handle::ServiceStateHandle;
 use overwatch_rs::services::relay::RelayMessage;
 use overwatch_rs::services::state::{NoOperator, NoState};
 use overwatch_rs::services::{ServiceCore, ServiceData, ServiceId};
 use overwatch_rs::DynError;
+use services_utils::overwatch::lifecycle;
 use std::fmt::{Debug, Formatter};
 use std::future::Future;
 use std::pin::Pin;
