+++ nomos-services/network/Cargo.toml
@@ -23,7 +23,7 @@ futures = "0.3"
 parking_lot = "0.12"
 nomos-core = { path = "../../nomos-core/chain-defs" }
 nomos-libp2p = { path = "../../nomos-libp2p", optional = true }
-nomos-utils = { path = "../../nomos-utils" }
+services-utils = { path = "../utils" }
 
 utoipa = { version = "4.0", optional = true }
 serde_json = { version = "1", optional = true }
