+++ nomos-services/system-sig/Cargo.toml
@@ -10,7 +10,7 @@ license = { workspace = true }
 async-trait = "0.1"
 async-ctrlc = "1.2"
 overwatch-rs = { git = "https://github.com/logos-co/Overwatch", rev = "f5f7ea0" }
-nomos-utils = { path = "../../nomos-utils" }
+services-utils = { path = "../utils" }
 tokio = "1"
 log = "0.4.20"
 futures = "0.3.28"
