+++ nomos-services/utils/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "utils"
+name = "services-utils"
 version = "0.1.0"
 edition = "2021"
 license = { workspace = true }
@@ -12,6 +12,4 @@ serde_json = "1.0.133"
 thiserror = "1"
 log = "0.4.22"
 async-trait = "0.1.83"
-
-[dev-dependencies]
 tracing = "0.1"
