+++ nomos-services/utils/src/overwatch/mod.rs
@@ -1,3 +1,4 @@
+pub mod lifecycle;
 pub mod recovery;
 
 pub use recovery::{FileBackend, JsonFileBackend, RecoveryError, RecoveryOperator};
