+++ nomos-utils/Cargo.toml
@@ -9,8 +9,6 @@ serde = ["dep:serde"]
 
 [dependencies]
 const-hex = "1"
-overwatch-rs = { git = "https://github.com/logos-co/Overwatch", rev = "f5f7ea0" }
 serde = { version = "1.0", optional = true }
 rand = "0.8"
 rand_chacha = "0.3"
-tracing = "0.1"
