+++ nomos-utils/src/lib.rs
@@ -1,5 +1,4 @@
 pub mod fisheryates;
-pub mod lifecycle;
 
 #[cfg(feature = "serde")]
 pub mod serde {
